Rename S3 settings section to S3 Bucket
Rename S3.vue to S3Bucket.vue, update the section id/label, switch
to lucide-archive icon, and reorder settings tabs to Git Settings,
S3 Bucket, Workers, Firewall.
